refactor vertical search acceptance fixtures into multiple files
Refactors the vertical search acceptance fixtures into multiple files
instead of sticking them all inside of one.

The facets fixtures are more complicated and stuffing them into one file
would be untenable.

J=SLAP-2151
TEST=none
