+++ tests/acceptance/fixtures/responses/vertical/KM/nyData.js
@@ -1,0 +1,109 @@
+export const nyData = {
+  data: {
+    id: '637478382857487577',
+    type: 'location',
+    website: 'https://locations.yext.com/us/ny/new-york/1-madison-ave.html',
+    address: {
+      line1: '1 Madison Ave',
+      line2: '5th Floor',
+      city: 'New York',
+      region: 'NY',
+      postalCode: '10010',
+      countryCode: 'US'
+    },
+    addressHidden: false,
+    description: 'Yext is the global digital knowledge management leader, and our Knowledge Engine puts businesses in control of their digital knowledge.',
+    hours: {
+      monday: {
+        openIntervals: [
+          {
+            start: '01:00',
+            end: '14:00'
+          }
+        ]
+      },
+      tuesday: {
+        openIntervals: [
+          {
+            start: '03:00',
+            end: '16:00'
+          }
+        ]
+      },
+      wednesday: {
+        openIntervals: [
+          {
+            start: '04:00',
+            end: '17:00'
+          },
+          {
+            start: '18:00',
+            end: '19:00'
+          }
+        ]
+      },
+      thursday: {
+        isClosed: true
+      },
+      friday: {
+        isClosed: true
+      },
+      saturday: {
+        isClosed: true
+      },
+      sunday: {
+        isClosed: true
+      }
+    },
+    name: 'NY Office Space',
+    cityCoordinate: {
+      latitude: 40.708601,
+      longitude: -73.876717
+    },
+    c_features: [
+      'Open Now',
+      'Dog Friendly'
+    ],
+    displayCoordinate: {
+      latitude: 40.7410895,
+      longitude: -73.9875092
+    },
+    geocodedCoordinate: {
+      latitude: 40.7410895,
+      longitude: -73.98750919999999
+    },
+    isoRegionCode: 'NY',
+    localPhone: '+12128136543',
+    mainPhone: '+18884442988',
+    priceRange: '$$',
+    routableCoordinate: {
+      latitude: 40.7411641,
+      longitude: -73.9878306
+    },
+    services: [
+      'Dogs',
+      'Cats',
+      'Sleep'
+    ],
+    timezone: 'America/New_York',
+    websiteUrl: {
+      url: 'https://locations.yext.com/us/ny/new-york/1-madison-ave.html',
+      preferDisplayUrl: false
+    },
+    yextDisplayCoordinate: {
+      latitude: 40.7410895,
+      longitude: -73.9875092
+    },
+    yextRoutableCoordinate: {
+      latitude: 40.7411641,
+      longitude: -73.9878306
+    },
+    categoryIds: [
+      '668'
+    ],
+    timeZoneUtcOffset: '-04:00',
+    uid: '18714997'
+  },
+  highlightedFields: {},
+  distance: 333778
+};
