+++ tests/acceptance/fixtures/responses/vertical/KM/ukData.js
@@ -1,0 +1,93 @@
+export const ukData = {
+  data: {
+    id: '7941089580646240971',
+    type: 'location',
+    website: 'https://locations.yext.com/gb/gt-lon/london/48-warwick-st-.html',
+    address: {
+      line1: '48 Warwick St.',
+      line2: 'Office 410',
+      city: 'London',
+      region: 'Gt Lon',
+      postalCode: 'W1B 5AW',
+      countryCode: 'GB'
+    },
+    addressHidden: false,
+    description: 'Yext is the global digital knowledge management leader, and our Knowledge Engine puts businesses in control of their digital knowledge.',
+    hours: {
+      monday: {
+        isClosed: true
+      },
+      tuesday: {
+        openIntervals: [
+          {
+            start: '15:00',
+            end: '16:00'
+          }
+        ]
+      },
+      wednesday: {
+        isClosed: true
+      },
+      thursday: {
+        isClosed: true
+      },
+      friday: {
+        isClosed: true
+      },
+      saturday: {
+        isClosed: true
+      },
+      sunday: {
+        isClosed: true
+      }
+    },
+    name: 'UK Office Space',
+    cityCoordinate: {
+      latitude: 51.50642013549805,
+      longitude: -0.12721000611782074
+    },
+    c_features: [
+      'Open Now'
+    ],
+    displayCoordinate: {
+      latitude: 51.5107139,
+      longitude: -0.1377914
+    },
+    geocodedCoordinate: {
+      latitude: 51.5106898,
+      longitude: -0.1378107
+    },
+    isoRegionCode: 'LND',
+    mainPhone: '+442037052290',
+    priceRange: '$$$$',
+    routableCoordinate: {
+      latitude: 51.51076,
+      longitude: -0.13767
+    },
+    services: [
+      'Dogs',
+      'Cats',
+      'Sleep'
+    ],
+    timezone: 'Europe/London',
+    websiteUrl: {
+      url: 'https://locations.yext.com/gb/gt-lon/london/48-warwick-st-.html',
+      preferDisplayUrl: false
+    },
+    yextDisplayCoordinate: {
+      latitude: 51.5107139,
+      longitude: -0.1377914
+    },
+    yextRoutableCoordinate: {
+      latitude: 51.51076,
+      longitude: -0.13767
+    },
+    categoryIds: [
+      '668'
+    ],
+    timeZoneUtcOffset: '+01:00',
+    uid: '18714999'
+  },
+  highlightedFields: {},
+  distance: 5900224
+};
