+++ tests/acceptance/fixtures/responses/vertical/KM/vaData.js
@@ -1,0 +1,119 @@
+export const vaData = {
+  data: {
+    id: '2917513710699998040',
+    type: 'location',
+    website: 'https://locations.yext.com/us/va/mclean/7900-westpark-drive.html',
+    address: {
+      line1: '7900 Westpark Drive',
+      line2: 'Suite T200',
+      city: 'McLean',
+      region: 'VA',
+      postalCode: '22102',
+      countryCode: 'US'
+    },
+    addressHidden: false,
+    description: 'Yext is the global digital knowledge management leader, and our Knowledge Engine puts businesses in control of their digital knowledge.',
+    hours: {
+      monday: {
+        openIntervals: [
+          {
+            start: '01:00',
+            end: '14:00'
+          }
+        ]
+      },
+      tuesday: {
+        openIntervals: [
+          {
+            start: '02:00',
+            end: '15:00'
+          },
+          {
+            start: '16:00',
+            end: '17:00'
+          }
+        ]
+      },
+      wednesday: {
+        openIntervals: [
+          {
+            start: '00:00',
+            end: '23:59'
+          }
+        ]
+      },
+      thursday: {
+        isClosed: true
+      },
+      friday: {
+        openIntervals: [
+          {
+            start: '12:00',
+            end: '14:00'
+          }
+        ]
+      },
+      saturday: {
+        openIntervals: [
+          {
+            start: '12:00',
+            end: '13:00'
+          }
+        ]
+      },
+      sunday: {
+        isClosed: true
+      }
+    },
+    name: 'VA Office Space',
+    cityCoordinate: {
+      latitude: 38.936519622802734,
+      longitude: -77.18428039550781
+    },
+    c_names: [
+      'my name is Steve',
+      'your name is Rose'
+    ],
+    displayCoordinate: {
+      latitude: 38.9246498,
+      longitude: -77.2169181
+    },
+    facebookPageUrl: 'https://www.facebook.com/Yext-Cafe-2073644659328705/',
+    geocodedCoordinate: {
+      latitude: 38.9246498,
+      longitude: -77.2169181
+    },
+    isoRegionCode: 'VA',
+    localPhone: '+16467624579',
+    mainPhone: '+18884442988',
+    routableCoordinate: {
+      latitude: 38.9243984,
+      longitude: -77.2178386
+    },
+    services: [
+      'Dogs',
+      'Cats',
+      'Sleep'
+    ],
+    timezone: 'America/New_York',
+    websiteUrl: {
+      url: 'https://locations.yext.com/us/va/mclean/7900-westpark-drive.html',
+      preferDisplayUrl: false
+    },
+    yextDisplayCoordinate: {
+      latitude: 38.9246498,
+      longitude: -77.2169181
+    },
+    yextRoutableCoordinate: {
+      latitude: 38.9243984,
+      longitude: -77.2178386
+    },
+    categoryIds: [
+      '668'
+    ],
+    timeZoneUtcOffset: '-04:00',
+    uid: '18714998'
+  },
+  highlightedFields: {},
+  distance: 13128
+};
