+++ tests/acceptance/fixtures/responses/vertical/people/amaniData.js
@@ -1,0 +1,88 @@
+export const amaniData = {
+  data: {
+    id: 'Employee-2116',
+    type: 'ce_person',
+    website: 'http://www.test.com',
+    address: {
+      line1: '7900 Westpark Drive',
+      city: 'Mclean',
+      region: 'VA',
+      postalCode: '22102',
+      countryCode: 'US'
+    },
+    associations: [
+      'Runners Association'
+    ],
+    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    name: 'Amani Farooque',
+    cityCoordinate: {
+      latitude: 38.936519622802734,
+      longitude: -77.18428039550781
+    },
+    c_allstateLadyCTA: {
+      label: 'Learn More',
+      linkType: 'URL',
+      link: 'http://yext.com'
+    },
+    c_employeeCity: 'Tysons Corner',
+    c_employeeCountry: 'United States',
+    c_employeeDepartment: 'Consulting',
+    c_employeeRegion: 'Virginia',
+    c_employeeTitle: 'Associate Technical Project Manager',
+    c_oliverCta: {
+      url: 'https://www.yext.com/s/2287528/entity/edit2?entityIds=13411251',
+      icon: 'yext',
+      label: 'test cta'
+    },
+    c_startDate: '2017-09-18',
+    displayCoordinate: {
+      latitude: 38.924648,
+      longitude: -77.216859
+    },
+    emails: [
+      'afarooque@yext.com',
+      'af@yext.com',
+      'aplomb@yext.com'
+    ],
+    firstName: 'Amani',
+    geocodedCoordinate: {
+      latitude: 38.924654,
+      longitude: -77.216891
+    },
+    headshot: {
+      url: 'https://a.mktgcdn.com/p/bmc3W88h2mMRXk-wHQ7dB0Em4rR_dfia6OVrAK3LjYU/192x191.png',
+      width: 192,
+      height: 191,
+      sourceUrl: 'http://a.mktgcdn.com/p/bmc3W88h2mMRXk-wHQ7dB0Em4rR_dfia6OVrAK3LjYU/192x191.png'
+    },
+    languages: [
+      'German'
+    ],
+    lastName: 'Farooque',
+    mainPhone: '+18003332222',
+    routableCoordinate: {
+      latitude: 38.9242966,
+      longitude: -77.2177549
+    },
+    specialities: [
+      'Documentation',
+      'Coding'
+    ],
+    websiteUrl: {
+      url: 'http://www.test.com',
+      displayUrl: 'http://www.testdisplay.com',
+      preferDisplayUrl: true
+    },
+    yextDisplayCoordinate: {
+      latitude: 38.924648,
+      longitude: -77.216859
+    },
+    yextRoutableCoordinate: {
+      latitude: 38.9242966,
+      longitude: -77.2177549
+    },
+    uid: '18716057'
+  },
+  highlightedFields: {},
+  distance: 13123
+};
