+++ tests/acceptance/fixtures/responses/vertical/people/tomData.js
@@ -1,0 +1,73 @@
+export const tomData = {
+  data: {
+    id: 'Employee-2143',
+    type: 'ce_person',
+    address: {
+      line1: '7900 Westpark Drive',
+      city: 'Mclean',
+      region: 'VA',
+      postalCode: '22102',
+      countryCode: 'US'
+    },
+    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+    name: 'Tom Meyer',
+    cityCoordinate: {
+      latitude: 38.936519622802734,
+      longitude: -77.18428039550781
+    },
+    c_allstateLadyCTA: {
+      label: 'Learn More',
+      linkType: 'URL',
+      link: 'http://yext.com'
+    },
+    c_employeeCity: 'Tysons Corner',
+    c_employeeCountry: 'United States',
+    c_employeeDepartment: 'Technology',
+    c_employeeRegion: 'Virginia',
+    c_employeeTitle: 'Software Engineer',
+    c_hierarchicalFacet: [
+      'Appliances',
+      'Appliances > Medium Appliances',
+      'Appliances > Medium Appliances > Cooking',
+      'Computer & Tablets',
+      'Computer & Tablets > Laptops'
+    ],
+    c_myRichTextField: '++Underlined stuff++  \n**Bold stuff**\n\n1. One\n2. Two\n3. *Three*',
+    c_popularity: '1.0',
+    c_startDate: '2017-09-05',
+    displayCoordinate: {
+      latitude: 38.924648,
+      longitude: -77.216859
+    },
+    emails: [
+      'tmeyer@yext.com'
+    ],
+    firstName: 'Tom',
+    geocodedCoordinate: {
+      latitude: 38.924654,
+      longitude: -77.216891
+    },
+    headshot: {
+      url: 'https://a.mktgcdn.com/p/cb9jR0A19ADuKc7ExlWmHaDchPW_61IR5TZ5G7NApjY/139x140.jpg',
+      width: 139,
+      height: 140,
+      sourceUrl: 'https://a.mktgcdn.com/p/cb9jR0A19ADuKc7ExlWmHaDchPW_61IR5TZ5G7NApjY/139x140.jpg'
+    },
+    lastName: 'Meyer',
+    routableCoordinate: {
+      latitude: 38.9242966,
+      longitude: -77.2177549
+    },
+    yextDisplayCoordinate: {
+      latitude: 38.924648,
+      longitude: -77.216859
+    },
+    yextRoutableCoordinate: {
+      latitude: 38.9242966,
+      longitude: -77.2177549
+    },
+    uid: '18716864'
+  },
+  highlightedFields: {},
+  distance: 13123
+};
