+++ tests/acceptance/fixtures/responses/vertical/sharedData.js
@@ -1,0 +1,20 @@
+export const meta = {
+  uuid: '018163fe-e697-6ffc-6346-d01618241911',
+  errors: []
+};
+
+export const basicResponseData = {
+  businessId: 3350634,
+  queryId: '018163fe-e6b4-af13-36e8-91d629a343d4',
+  results: [],
+  resultsCount: 0,
+  source: 'KNOWLEDGE_MANAGER',
+  searchIntents: []
+};
+
+export const locationBias = {
+  latitude: 38.890396,
+  longitude: -77.084159,
+  locationDisplayName: 'Arlington, Virginia, United States',
+  accuracy: 'DEVICE'
+};
